docs/archive/teanga/escriptorium/app/requirements.txt
albumentations
beautifulsoup4~=4.7.1
bleach~=5.0.1
celery~=5.3.6
channels~=3.0.4
channels-redis~=3.4.0
daphne~=3.0.2
Django~=4.2.13
django-cleanup~=8.1.0
django-country-and-language-flags
django-filter
django-ordered-model~=3.7
django-prometheus~=2.3.1
django-redis~=5.4.0
django-simple-captcha~=0.6.0
django-solo>=1.1
djangorestframework~=3.15.1
drf-nested-routers~=0.91
easy-thumbnails~=2.8.1
elasticsearch~=7.17.0
kraken>=5.2.9,~=5.2
oitei~=1.1.1
git+https://github.com/dasmiq/passim.git@v2.0.1#egg=passim
Pillow>=5.4.1
psutil
psycopg2-binary
pyvips~=2.1.12
redis~=4.4.4
requests>=2.21
sentry-sdk>=1.10
uWSGI~=2.0.20
